Moving sub components to the parent
