+++ lib/components/Keyboard/KeyboardAccessoryView/index.tsx
@@ -10,8 +10,8 @@ import {
   LayoutChangeEvent
 } from 'react-native';
 import KeyboardTrackingView, {KeyboardTrackingViewProps} from '../KeyboardTrackingView';
-import CustomKeyboardView from '../CustomKeyboardView';
-import KeyboardUtils, {KeyboardHeightListener} from '../KeyboardUtils';
+import CustomKeyboardView from './CustomKeyboardView';
+import KeyboardUtils, {KeyboardHeightListener} from './KeyboardUtils';
 
 const IsIOS = Platform.OS === 'ios';
 const IsAndroid = Platform.OS === 'android';
