+++ lib/components/Keyboard/index.ts
@@ -1,8 +1,8 @@
 import KeyboardTrackingView, {KeyboardTrackingViewProps} from './KeyboardTrackingView';
 import KeyboardAwareInsetsView from './KeyboardAwareInsetsView';
 import KeyboardAccessoryView, {KeyboardAccessoryViewProps} from './KeyboardAccessoryView';
-import KeyboardRegistry from './KeyboardRegistry';
-import KeyboardUtils from './KeyboardUtils';
+import KeyboardRegistry from './KeyboardAccessoryView/KeyboardRegistry';
+import KeyboardUtils from './KeyboardAccessoryView/KeyboardUtils';
 
 export {KeyboardTrackingViewProps, KeyboardAccessoryViewProps};
